refactor(frontend): split LoopKanban into sub-components
Split LoopKanban (615→1 line) into:
- loop-kanban/helpers.tsx: execStatusView, durationLabel, COLUMNS
- loop-kanban/ExecutionCard.tsx: execution card with sub-components
- loop-kanban/KanbanColumn.tsx: kanban column
- loop-kanban/useLoopExecutions.ts: data fetching hook
- loop-kanban/index.tsx: main component

All exports maintained via re-export from original file.
